extract fetch reservation function to a hook
